+++ src/v2019C1/dialecte.ts
@@ -1,16 +1,18 @@
 import { SCL_DIALECTE_CONFIG } from './config/dialecte.config'
-import { EXTENSIONS } from './extensions'
+import { SCL_EXTENSION_MODULES } from './extensions'
 
 import { openDialecteDocument } from '@dialecte/core'
 
-import type { StorageOptions, Document } from '@dialecte/core'
+import type { StorageOptions, ExtensionModules } from '@dialecte/core'
 
-export function openSclDocument(
-	storage: StorageOptions,
-): Document<typeof SCL_DIALECTE_CONFIG, typeof EXTENSIONS> {
+export function openSclDocument<
+	CustomModules extends ExtensionModules = Record<never, never>,
+>(params: { storage: StorageOptions; extensions?: CustomModules }) {
+	const { storage, extensions } = params
+
 	return openDialecteDocument({
 		config: SCL_DIALECTE_CONFIG,
 		storage,
-		extensions: EXTENSIONS,
+		extensions: { base: SCL_EXTENSION_MODULES, custom: extensions },
 	})
 }