+++ src/v2019C1/extensions/history/query/get-latest-hitem.test.ts
@@ -82,5 +82,5 @@ describe('getLatestHitem', () => {
 		return { assertDatabaseName: source.databaseName }
 	}
 
-	runSclTestCases({ testCases, act })
+	runSclTestCases.withExport({ testCases, act })
 })