+++ src/v2019C1/extensions/history/query/get-sorted-hitem.test.ts
@@ -83,5 +83,5 @@ describe('getSortedHitems', () => {
 		return { assertDatabaseName: source.databaseName }
 	}
 
-	runSclTestCases({ testCases, act })
+	runSclTestCases.withExport({ testCases, act })
 })