+++ src/v2019C1/extensions/history/transaction/add-history-entry.test.ts
@@ -110,5 +110,5 @@ describe('addHistoryEntry', () => {
 		return { assertDatabaseName: source.databaseName }
 	}
 
-	runSclTestCases({ testCases, act })
+	runSclTestCases.withExport({ testCases, act })
 })