+++ src/v2019C1/extensions/index.ts
@@ -2,6 +2,4 @@ import { dataModel } from './data-model'
 import { history } from './history'
 import { template } from './template'
 
-import { mergeExtensions } from '@dialecte/core/helpers'
-
-export const EXTENSIONS = mergeExtensions({ history, dataModel, template })
+export const SCL_EXTENSION_MODULES = { history, dataModel, template }