+++ src/v2019C1/extensions/template/ensure-substation-structure.test.ts
@@ -50,7 +50,7 @@ describe('ensureSubstationStructure', () => {
 		},
 	}
 
-	runSclTestCases({
+	runSclTestCases.withExport({
 		testCases,
 		act: async ({ source }) => {
 			await source.document.transaction(async (tx) => {