+++ src/v2019C1/io/export.ts
@@ -3,7 +3,6 @@ import { SCL_DIALECTE_CONFIG } from '../config'
 import { exportXmlFile } from '@dialecte/core'
 
 import type { Config } from '@/v2019C1/config'
-import type { Scl } from '@/v2019C1/config'
 
 export function exportSclFile(params: {
 	databaseName: string